Abstract Data Type str a simple string class
Purpose:
This package illustrates:
1. How to lay out a package
2. Construction of a relatively simple abstract data type
I like the way this class works, the language it provides, and the range of its operations.
However, I do not use the class myself, and don’t suggest you use it.
Prefer the std::string class.  Everyone knows that and how it works.
Str is intended for instructional purposes only.